kube-gen: fixup moved tests
