--- staging/src/k8s.io/kube-gen/cmd/client-gen/.import-restrictions
@@ -1,1 +1,0 @@
-{}
