+++ staging/src/k8s.io/kube-gen/test/apis/testgroup/doc.go
@@ -16,4 +16,4 @@ limitations under the License.
 
 // +k8s:deepcopy-gen=package,register
 // +groupName=testgroup.k8s.io
-package testgroup // import "k8s.io/kube-gen/cmd/client-gen/test_apis/testgroup"
+package testgroup // import "k8s.io/kube-gen/test/apis/testgroup"
