+++ staging/src/k8s.io/kube-gen/test/apis/testgroup/install/install.go
@@ -22,8 +22,8 @@ import (
 	"k8s.io/apimachinery/pkg/apimachinery/announced"
 	"k8s.io/apimachinery/pkg/apimachinery/registered"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/kube-gen/cmd/client-gen/test_apis/testgroup"
-	"k8s.io/kube-gen/cmd/client-gen/test_apis/testgroup/v1"
+	"k8s.io/kube-gen/test/apis/testgroup"
+	"k8s.io/kube-gen/test/apis/testgroup/v1"
 )
 
 // Install registers the API group and adds types to a scheme
@@ -32,7 +32,7 @@ func Install(groupFactoryRegistry announced.APIGroupFactoryRegistry, registry *r
 		&announced.GroupMetaFactoryArgs{
 			GroupName:                  testgroup.SchemeGroupVersion.Group,
 			VersionPreferenceOrder:     []string{v1.SchemeGroupVersion.Version},
-			ImportPrefix:               "k8s.io/kube-gen/cmd/client-gen/test_apis/testgroup",
+			ImportPrefix:               "k8s.io/kube-gen/test/apis/testgroup",
 			AddInternalObjectsToScheme: testgroup.AddToScheme,
 		},
 		announced.VersionToSchemeFunc{
