+++ staging/src/k8s.io/kube-gen/test/clientset/internal/doc.go
@@ -17,4 +17,4 @@ limitations under the License.
 // This package is generated by client-gen with custom arguments.
 
 // This package has the automatically generated clientset.
-package test_internalclientset
+package internal
